NINETEEN HOURS **STRAIGHT** OF PAINFUL PIEN 🥺

+++ src/main/java/com/hbm/util/I18nUtil.java
@@ -13,7 +13,7 @@ public class I18nUtil {
 	}
 
 	public static class leafia {
-		public static String[] statusDecimals(String template,double value,byte decimals) {
+		public static String[] statusDecimals(String template,double value,int decimals) {
 			double mul = Math.pow(10,decimals);
 			return resolveKeyArray(template,String.format("%01."+decimals+"f",Math.floor(value*mul+0.5)/mul));
 		}
